# services/manager/openapi/v1.yaml
openapi: 3.1.0
info:
  title: SkausWatch Manager API
  description: SkausWatch manager service — users, alerts, threat intelligence, approvals, ENDPOINT agent fleet, SIEM, ASM, CodeScan proxy, S3 malware scanning, and research lookups.
  contact:
    name: Penguin Tech Inc
    email: support@penguintech.io
  license:
    name: AGPL-3.0-only
    identifier: AGPL-3.0-only
  version: '1'
paths:
...
  /api/v1/auth/refresh:
    post:
      tags:
      - auth
      summary: |-
        POST /auth/refresh — unauthenticated (the refresh token in the body is
        the credential; no bearer header is required or checked).
      operationId: refresh
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RefreshRequest'
        required: true
      responses:
        '200':
          description: Rotated access/refresh token pair
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RefreshResponse'
        '401':
          description: Invalid, revoked, or expired refresh token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      description: |-
        Documentation-only mirror of `ApiError`'s bare `{"error": msg}` wire shape
        (`BadRequest`/`Unauthorized`/`Forbidden`/`NotFound`/`Conflict` in this
        service all use it, when their body happens to be bare) — `ApiError`
        itself builds its body with `serde_json::json!` rather than a typed
        struct, so this type exists solely to give `utoipa` something to
        reference in `#[utoipa::path]` `responses(...)` clauses. See
        `docs/v2-port/openapi-pattern.md`.
      required:
      - error
      properties:
        error:
          type: string
          description: Human-readable error message.
...
    RefreshRequest:
      type: object
      description: POST /auth/refresh body.
      required:
      - refresh_token
      properties:
        refresh_token:
          type: string
    RefreshResponse:
      type: object
      description: Documentation-only mirror of `refresh`'s `serde_json::json!` body.
      required:
      - access_token
      - refresh_token
      - token_type
      - expires_in
      properties:
        access_token:
          type: string
        expires_in:
          type: integer
          format: int64
          description: Access-token lifetime in seconds.
        refresh_token:
          type: string
        token_type:
          type: string
...
  securitySchemes:
    bearer_jwt:
      type: http
      scheme: bearer
      bearerFormat: JWT
    endpoint_hmac:
      type: apiKey
      in: header
      name: X-API-Key
...